# Parallelize secondary cache lookup in MultiGet